feat: add validation for delete functionality and create UpdateItemIsEmptyException
test: add should_response_404_when_delete_given_invalid_id test for TodoListBackendIntegrationTests

+++ src/main/java/oocl/todolistbackend/service/TodoService.java
@@ -40,7 +40,9 @@ public class TodoService {
     }
 
     public void delete(long id) {
-
+        if(!todoRepository.getTodoById(id)){
+            throw new TodoItemNotFoundException();
+        }
         todoRepository.remove(id);
     }
 }
